+++ data/fs/files/products.json
@@ -1,6 +1,6 @@
 [
   {
-    "id": "692c34096fa235b63eb9172e",
+    "id": "1f8d33626cd4099e1e01c3a9",
     "title": "Sofá de Cuero",
     "photo": "sofa_cuero.jpg",
     "category": "Muebles",
@@ -9,7 +9,16 @@
     "stock": 1
   },
   {
-    "id": "1d0618d0b085317a859da5c9",
+    "id": "dd483ca53281b05cc2f7963c",
+    "title": "Mesa de Comedor de Madera",
+    "photo": "mesa_comedor.jpg",
+    "category": "Muebles",
+    "description": "Mesa de comedor de madera maciza con capacidad para 6 personas. Estilo rústico.",
+    "price": 250,
+    "stock": 1
+  },
+  {
+    "id": "744bd377cbd83bf41d69c227",
     "title": "Televisor LED Samsung 55 Pulgadas",
     "photo": "tv_samsung.jpg",
     "category": "Electrónica",
@@ -18,7 +27,7 @@
     "stock": 1
   },
   {
-    "id": "d33d7a296939cfaa0c23838d",
+    "id": "a26e8dd2741d2baf55fc9adc",
     "title": "Cámara Réflex Nikon D5600",
     "photo": "camara_nikon.jpg",
     "category": "Electrónica",
@@ -27,7 +36,7 @@
     "stock": 1
   },
   {
-    "id": "469a7baccf0ebf1bb536bf7c",
+    "id": "181f32d6527b76a54361dd98",
     "title": "Mesa de Centro de Vidrio",
     "photo": "mesa_centro.jpg",
     "category": "Muebles",
@@ -36,7 +45,7 @@
     "stock": 1
   },
   {
-    "id": "16dc7b05ed2bcb25ae5968a9",
+    "id": "081546f470d58b7805909a60",
     "title": "Batidora de Mano Bosch",
     "photo": "batidora_bosch.jpg",
     "category": "Electrodomésticos",
@@ -45,7 +54,7 @@
     "stock": 1
   },
   {
-    "id": "379469b8d26f4af4bb1f1124",
+    "id": "74d56dbf88834d1b969bd63a",
     "title": "Robot Aspirador Roomba 960",
     "photo": "roomba_960.jpg",
     "category": "Electrodomésticos",
@@ -54,7 +63,7 @@
     "stock": 1
   },
   {
-    "id": "7dd26c40e58123f9ffc82f5f",
+    "id": "bc666326c7e47fd8b23bd473",
     "title": "Silla de Oficina Ergonómica",
     "photo": "silla_oficina.jpg",
     "category": "Muebles",
@@ -63,7 +72,7 @@
     "stock": 1
   },
   {
-    "id": "e36bd05d47c1b343bd9e8829",
+    "id": "519c8732494e1e292dce8a4e",
     "title": "Licuadora Oster de Alta Velocidad",
     "photo": "licuadora_oster.jpg",
     "category": "Electrodomésticos",
@@ -72,12 +81,102 @@
     "stock": 1
   },
   {
-    "id": "f449edfac4d319ec1f70259a",
+    "id": "797069df8850c5a1df4e2f36",
     "title": "Mesa de Noche de Roble",
     "photo": "mesa_noche.jpg",
     "category": "Muebles",
     "description": "Mesa de noche de roble con cajón y estante inferior. Perfecta para dormitorios modernos.",
     "price": 80,
+    "stock": 1
+  },
+  {
+    "id": "54c01484918512a68581b62d",
+    "title": "Aire Acondicionado Inverter LG 12000 BTU",
+    "photo": "aire_acondicionado.jpg",
+    "category": "Electrodomésticos",
+    "description": "Aire acondicionado LG con tecnología Inverter. Eficiente y silencioso. Ideal para habitaciones grandes.",
+    "price": 800,
+    "stock": 1
+  },
+  {
+    "id": "d0e8439c5ddd2d6bd2e2fcda",
+    "title": "Lavadora Samsung Carga Frontal 9 Kg",
+    "photo": "lavadora_samsung.jpg",
+    "category": "Electrodomésticos",
+    "description": "Lavadora Samsung de carga frontal con capacidad para 9 kg. Programas de lavado personalizables.",
+    "price": 600,
+    "stock": 1
+  },
+  {
+    "id": "d33e066af91202446d2c8a16",
+    "title": "Sofá Cama Reversible con Almacenamiento",
+    "photo": "sofa_cama.jpg",
+    "category": "Muebles",
+    "description": "Sofá cama reversible con almacenamiento. Perfecto para espacios pequeños. Disponible en varios colores.",
+    "price": 450,
+    "stock": 1
+  },
+  {
+    "id": "c0a3ac9b56b867f63bdee5b8",
+    "title": "Televisor OLED LG 65 Pulgadas 4K",
+    "photo": "tv_lg_oled.jpg",
+    "category": "Electrónica",
+    "description": "Televisor OLED LG de 65 pulgadas con resolución 4K. Colores vibrantes y negros perfectos.",
+    "price": 1500,
+    "stock": 1
+  },
+  {
+    "id": "f11ffb6f6ddd379cfbbbd57d",
+    "title": "Mesa de Escritorio de Estilo Industrial",
+    "photo": "escritorio_industrial.jpg",
+    "category": "Muebles",
+    "description": "Mesa de escritorio de estilo industrial con patas de metal y tablero de madera. Diseño moderno y funcional.",
+    "price": 200,
+    "stock": 1
+  },
+  {
+    "id": "f1f62d1845e32f7bb5a90a37",
+    "title": "Cafetera Expresso De'Longhi",
+    "photo": "cafetera_delonghi.jpg",
+    "category": "Electrodomésticos",
+    "description": "Cafetera Expresso De'Longhi con sistema de espuma de leche. Prepara café como un barista profesional.",
+    "price": 300,
+    "stock": 1
+  },
+  {
+    "id": "e133e8671c243767b20a62a2",
+    "title": "Colchón Viscoelástico con Gel",
+    "photo": "colchon_viscoelastico.jpg",
+    "category": "Hogar",
+    "description": "Colchón viscoelástico con gel que se adapta a la forma del cuerpo. Descanso óptimo y confortable.",
+    "price": 700,
+    "stock": 1
+  },
+  {
+    "id": "bfb330ab4e025cacf321d128",
+    "title": "Sistema de Sonido Envolvente 5.1",
+    "photo": "sistema_sonido_51.jpg",
+    "category": "Electrónica",
+    "description": "Sistema de sonido envolvente 5.1 con altavoces inalámbricos. Sumérgete en la experiencia del cine en casa.",
+    "price": 400,
+    "stock": 1
+  },
+  {
+    "id": "25c2225813898a86adbb47c9",
+    "title": "Juego de Comedor de Ratán Sintético",
+    "photo": "comedor_ratan.jpg",
+    "category": "Muebles",
+    "description": "Juego de comedor de ratán sintético para exteriores. Resistente a la intemperie y fácil de limpiar.",
+    "price": 600,
+    "stock": 1
+  },
+  {
+    "id": "55950f7a35648328cfb6f088",
+    "title": "Máquina de Ejercicio Multifuncional",
+    "photo": "maquina_ejercicio.jpg",
+    "category": "Deportes y Fitness",
+    "description": "Máquina de ejercicio multifuncional para entrenamiento en casa. Tonifica y fortalece todos los grupos musculares.",
+    "price": 800,
     "stock": 1
   }
 ]
